Added VMS support for mx_cfn.c and mx_cpu_arch.c
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@392 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_record.c
@@ -1779,7 +1779,7 @@ mx_print_structure( FILE *file, MX_RECORD *record, unsigned long mask )
 					"ERROR: Field %lu is NULL!!!\n", i );
 			}
 
-			if ( mask == MXFF_SHOW_ALL ) {
+			if ( mask == (unsigned long) MXFF_SHOW_ALL ) {
 				show_field = TRUE;
 			} else {
 				if ( field->flags & mask ) {
